requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
openai
chromadb
python-dotenv
